+++ include/ae/ae.h
@@ -9,6 +9,7 @@
 #ifndef _AE_H
 #define _AE_H
 
+#include <ae/event.h>
 #include <ae/io.h>
 #include <ae/log.h>
 #include <ae/misc.h>