+++ CHANGELOG.md
@@ -7,6 +7,10 @@ project adheres
 to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED] 2018-03-12
+### Added 
+- ae_event_t to wrap eventfd
+- ae_pool_reset()
+
 ### Fixed
 - io.h was missing from the install target in Makefile.am